publication and paragraph files added

+++ src/main/java/com/example/news/entity/Publication.java
@@ -18,7 +18,7 @@ import java.util.Set;
 @Entity
 public class Publication extends Auditable {
     @Column(name = "count_views")
-    private int countViews;
+    private Long countViews;
 
     @ManyToOne
     @JoinColumn(name="category_id", foreignKey = @ForeignKey(name="FK_CATEGORY__ID"), nullable=false)
